requirements.txt
openai-agents
Flask
pytest-playwright